# [MRG] ENH: Improves speed of one hot encoding