This PR improves the time it takes to run one hot encoding by about 24%.

### Benchmark script

```py
from itertools import product
import neurtu
from sklearn.preprocessing import OneHotEncoder
import numpy as np


def cases():
    rng = np.random.RandomState(42)

    n_features = [50, 100]
    n_samples = [100_000, 200_000, 500_000]

    for n_samples, n_features in product(n_samples, n_features):
        X = rng.randint(0, 75, size=(n_samples, n_features))
        tags = {'n_samples': n_samples, 'n_features': n_features}
        enc = OneHotEncoder(categories='auto')
        yield neurtu.delayed(enc, tags=tags).fit_transform(X)


bench = neurtu.Benchmark(wall_time=True, cpu_time=True, to_dataframe=True)
df = bench(cases())
print(df)
```

### This PR

```
                      wall_time  cpu_time
n_samples n_features
100000    50           0.531036  1.808001
          100          1.179558  2.429228
200000    50           1.355601  2.621355
          100          2.950725  4.143948
500000    50           3.646070  4.799419
          100          7.739459  8.745608
```

### Master

```
                      wall_time   cpu_time
n_samples n_features
100000    50           0.728843   2.058129
          100          1.598936   3.000086
200000    50           1.825290   3.265459
          100          3.882629   5.505839
500000    50           4.915225   6.444236
          100         10.473135  12.773251
```

CC @jorisvandenbossche